add DoubleField

+++ random-generator/src/main/java/com/hormones/random/field/range/DateTimeField.java
@@ -22,7 +22,7 @@ public class DateTimeField extends RangeField<LocalDateTime> {
 
     @Override
     protected LocalDateTime getIncrement() {
-        return this.value.plusSeconds(this.auto);
+        return this.value.plusSeconds(this.auto.longValue());
     }
 
     @Override
